Use 32 bit count types.

+++ include/nois/NoisTypes.hpp
@@ -6,8 +6,8 @@
 
 namespace nois {
 
-using count_t = int64_t;
-using ucount_t = uint64_t;
+using count_t = int32_t;
+using ucount_t = uint32_t;
 using f32_t = float;
 using f64_t = double;
 using s32_t = int32_t;
